+++ mullvad-daemon/src/exception_logging/mod.rs
@@ -1,0 +1,11 @@
+#[cfg(windows)]
+mod win;
+
+#[cfg(windows)]
+pub use win::enable;
+
+#[cfg(unix)]
+mod unix;
+
+#[cfg(unix)]
+pub use unix::enable;